Fixed a bug that made Calendar2_Helper_Rrule crash on empty INTERVALs
The helper checks with is_null if the INTERVAL is empty. Sadly, it did this
_after_ casting the value to int, which is never true, as (int) null is
evaluated as 0.

+++ phprojekt/application/Calendar2/Helper/Rrule.php
@@ -234,13 +234,13 @@ class Calendar2_Helper_Rrule
 
     private static function _parseInterval($rrule)
     {
-        $interval = (int) self::_extractFromRrule($rrule, 'INTERVAL');
+        $interval = self::_extractFromRrule($rrule, 'INTERVAL');
         if (is_null($interval)) {
             $interval = 1;
         } else if (0 >= $interval) {
             throw new Exception('Negative or Zero Intervals not permitted.');
         }
-        return $interval;
+        return (int) $interval;
     }
 
     private static function _parseFreq($rrule)
